raspberry_circuit/requirements.txt
iniconfig==2.0.0
mypy==1.11.1
mypy-extensions==1.0.0
packaging==24.1
paho-mqtt==2.1.0
pluggy==1.5.0
pytest==8.3.2
pytest-mock==3.14.0
typing_extensions==4.12.2
